tinyalsa: Add a board config to drop silence_size for blob compat.
In cm-13.0, this commit was added:

commit e25fe0b50ea717a9f347dcb927d396e8772e6362
Author: Maneet Singh <mmaneetsingh@nvidia.com>
Date:   Thu Jun 11 17:34:40 2015 -0700

    pcm: add support to set silence_size

    Tinyalsa always set silence_size to zero. Add support to set this
    pcm software parameter as required.

    Bug: 20226809
    Bug: 20300203
    Change-Id: I25de43623dc04bf5a3ad4c6573bc2b8dad1eec5e

that changes the size of the pcm_config structure.  This is a
problem for any device using older blobs that call pcm_open
as their structure is now invalid.

Fixing this via a prebuilt blob is not ideal because we have
source built components that also use tinyalsa (e.g. the qcom audio
HALs).  Using a blob (either as a complete replacement or shim)
would involve having code in the same address space that uses two
incompatible copies of pcm_open (e.g. via a proprietary audio.primary.
and a source built audio.usb.).

Therefore, offer a compile option to enable this backward compatibility
support.

Change-Id: I4d96f2b1fe841fc19e83ddf9a4ca6c0f04bcd3f5
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>